move phone-registration
