+++ lib/database.js
@@ -3,7 +3,7 @@ import { createHash, randomBytes, createCipheriv, createDecipheriv } from 'crypt
 import Listener from 'pg-listen';
 import logger from './logger.js';
 import { getHandler } from './handlers/index.js';
-import { initPhoneRegistration } from './models/phone-registration.js';
+import { initPhoneRegistration } from './database-models/phone-registration.js';
 import { encryptSecret, decryptSecret } from './utils/credentials.js';
 
 const schemaVersion = 17;
